pull out global premium index type and funding rate type
